Move Loader to shared components
