+++ src/components/shared/loader/loader.scss
